Finished upgrade system Work on Emplacements Removed unncecessary comments Machinegun skin text rendering works Added AABBs for almost all multiblocks Gas cloud fixes (thanks Carver) Naval mine fixes (thanks Carver)

+++ src/main/java/pl/pabilo8/immersiveintelligence/client/gui/GuiElectrolyzer.java
@@ -60,7 +60,7 @@ public class GuiElectrolyzer extends GuiIEContainerBase
 
 		if(tile.active&&tile.processTimeMax!=0)
 		{
-			float progress = Math.min(1f, tile.processTime/tile.processTimeMax);
+			float progress = Math.min(1f, tile.processTime/(float)tile.processTimeMax);
 			this.drawTexturedModalRect(guiLeft+66, guiTop+42, 176, 71, Math.round(54*progress), 10);
 		}
 	}
